WebCrawler/WebCrawler/requirements.txt
altgraph==0.17.2
appdirs==1.4.4
async-generator==1.10
attrs==21.4.0
audioread==3.0.0
certifi==2022.6.15
cffi==1.15.0
chardet==5.2.0
charset-normalizer==2.1.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==37.0.2
cycler==0.11.0
ddddocr==1.5.5
decorator==5.1.1
flatbuffers==24.3.25
fonttools==4.38.0
future==0.18.2
h11==0.13.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.12.0
inflect==5.3.0
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.4.4
librosa==0.8.1
llvmlite==0.39.1
lxml==4.9.2
matplotlib==3.5.1
mpmath==1.3.0
numba==0.56.4
numpy==1.21.6
onnxruntime==1.14.1
opencv-python-headless==4.10.0.84
outcome==1.2.0
packaging==21.3
pandas==1.3.5
pefile==2022.5.30
Pillow==8.4.0
pip==24.0
pooch==1.6.0
protobuf==4.24.4
psutil==5.9.6
pure-python-adb==0.3.0.dev0
pycparser==2.21
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2022.8
pyLSV2==0.7.7
pynndescent==0.5.8
pyodbc==4.0.32
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyreadline==2.1
PySocks==1.7.1
pytesseract==0.3.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
pywin32-ctypes==0.2.2
pyzmq==24.0.1
requests==2.28.1
resampy==0.4.2
scikit-learn==1.0.2
scipy==1.7.3
selenium==4.11.2
setuptools==47.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sounddevice==0.4.3
SoundFile==0.10.3.post1
sympy==1.10.1
tesseract==0.1.3
threadpoolctl==3.1.0
torchfile==0.1.0
tornado==6.2
tqdm==4.62.3
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.2.0
umap-learn==0.5.2
Unidecode==1.3.2
urllib3==1.26.7
visdom==0.1.8.9
webdriver-manager==4.0.0
webrtcvad==2.0.10
websocket-client==1.4.2
wsproto==1.1.0
zipp==3.8.1
